util.hmac: fix HMAC for SHA384 and SHA512
According to RFC4868, blocksize for SHA256 is 512 bits (64 bytes), but
for SHA384 and SHA512 it is 1024 bits (128 bytes).

Also, RFC4231 provides test vectors which fail when blocksize for SHA512
is 64 bytes.

qxmpp-project/qxmpp#177 mentions that the client fails to authenticate
with SCRAM-SHA-512. Libstrophe library fails to authenticate as well.
However, authentication is successful when libstrophe uses blocksize
64 for HMAC-SHA-512.